requirements.txt
streamlit
pdf2image
pytesseract
sentence-transformers
faiss-cpu
numpy
ollama
Pillow